fix symbolictensor conversion to int using native tf.cast

+++ mlu_tools/processing/image.py
@@ -88,7 +88,8 @@ def perform_oversampling(dir_pth, target_size):
 
 def process_path(file_path, image_size, crop_to_aspect_ratio):
     label = tf.strings.split(file_path, os.sep)[-2]
-    label = int(label)
+    # label = int(label)
+    label = tf.cast(label, dtype=tf.int32)
     img = tf.io.read_file(file_path)
     img = tf.image.decode_jpeg(img, channels=3)
     # img = tf.image.resize(img, image_size)  # doesn't support center cropping
